requirements.txt
PyExecJS
lzstring~=1.0.4
pillow~=9.1.0
bs4~=0.0.1
configparser~=5.2.0
PyQt5~=5.15.6
langcodes~=3.3.0
language_data
beautifulsoup4~=4.11.1
imagesize~=1.3.0

ebooklib~=0.17.1
python-docx
jsbeautifier~=1.14.3
pymupdf
fitz~=0.0.1.dev2
googletrans==4.0.0rc1

pytesseract~=0.3.9
opencv-python~=4.5.5.62
numpy~=1.22.3
requests~=2.27.1
docx~=0.2.4